+++ src/main/java/roomescape/reservation/controller/TimeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.reservation.controller;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,10 +9,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.time.TimeRequest;
-import roomescape.dto.time.TimeResponse;
-import roomescape.model.Time;
-import roomescape.model.TimeDao;
+import roomescape.reservation.controller.dto.TimeRequest;
+import roomescape.reservation.controller.dto.TimeResponse;
+import roomescape.reservation.domain.Time;
+import roomescape.reservation.domain.repository.TimeDao;
 
 @RequestMapping("/times")
 @RestController
