+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.reservation.domain;
 
 import java.time.LocalDate;
